ml: send & massSend

+++ functions/functions.php
@@ -126,12 +126,15 @@ function get_permissions($page) {
  * @return null
  */
 function _log ($msg, $log='all.log') {
-	if (trim(strtolower($msg))==='clear')
-		unlink ($log);
-	else if ($logfile = fopen($log, 'a')) {
-		fwrite($logfile, date('Y-m-d H:i:s - ').$msg."\n");
-		fclose($logfile);
-	}
+	global $app;
+	if ($app['log']):
+		if (trim(strtolower($msg))==='clear')
+			unlink ($log);
+		else if ($logfile = fopen($log, 'a')) {
+			fwrite($logfile, date('Y-m-d H:i:s - ').$msg."\n");
+			fclose($logfile);
+		}
+	endif;
 	return;
 }
